libfdt: Fix the undefined behavior in fdt_num_mem_rsv()
With LLVM 10.0.0+, the following codes in fdt_num_mem_rsv() does not
work any more for an fdt that is at address 0:

    for (i = 0; (re = fdt_mem_rsv(fdt, i)) != NULL; i++) {
        if (fdt64_ld_(&re->size) == 0)
            return i;
    }

Due to LLVM's optimization engine utilizing a UB in C, the following
code pattern:

    if ((pointer + offset) != NULL)

is transformed into:

    if (pointer != NULL)

because if pointer is NULL and offset is non-zero, the result of
(pointer + offset) is UB. So LLVM is free to exploit such UB to
perform some optimization.

In this case, fdt_mem_rsv() gets inlined and returns (pointer + offset).
And LLVM in turns emits codes to check fdt against NULL, which won't
work for fdt at address 0.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>